Improve configuration meta-data
If a nested group is defined by an inner class of the current
configuration class, there is no need to flag it with
`@NestedConfigurationProperty`. And there is not need to provide a setter
as long as the instance is available (i.e. making it final is a best
practice unless you want to lazily create it).

Also all keys were missing documentation which are self-explanatory here
but it doesn't hurt.